refactor: add support for reasoning message part in object view sidebar
