+++ src/content-view/object-sidebar-content.tsx
@@ -1,11 +1,9 @@
 import { SidebarContent } from '@/components/ui/sidebar'
-import { getToolMetadataSync } from '@/lib/tool-metadata'
-import { formatToolOutput, splitPartType } from '@/lib/utils'
-import type { ToolUIPart } from 'ai'
 import { ContentViewHeader } from './header'
+import { type ObjectViewData } from './context'
 
 type ObjectSidebarContentProps = {
-  content: ToolUIPart
+  content: ObjectViewData
   onClose: () => void
 }
 
@@ -13,16 +11,11 @@ type ObjectSidebarContentProps = {
  * Content for displaying tool call results in the unified content view
  */
 export const ObjectSidebarContent = ({ content, onClose }: ObjectSidebarContentProps) => {
-  const [, toolName] = splitPartType(content?.type ?? '')
-  const metadata = getToolMetadataSync(toolName, content?.input)
-
   return (
     <div className="flex flex-col h-full">
-      <ContentViewHeader title={metadata.displayName} onClose={onClose} className="bg-card border-b border-border" />
+      <ContentViewHeader title={content.title} onClose={onClose} className="bg-card border-b border-border" />
       <SidebarContent className="p-4 overflow-x-hidden">
-        <p className="text-gray-700 dark:text-gray-300 text-sm whitespace-pre-wrap">
-          {formatToolOutput(content.output)}
-        </p>
+        <p className="text-gray-700 dark:text-gray-300 text-sm whitespace-pre-wrap">{content.output}</p>
       </SidebarContent>
     </div>
   )
